+++ R/arkhe-package.R
@@ -2,7 +2,7 @@
 #'  \tabular{ll}{
 #'   **Package:** \tab arkhe \cr
 #'   **Type:** \tab Package \cr
-#'   **Version:** \tab 1.2.0 \cr
+#'   **Version:** \tab 1.3.0 \cr
 #'   **License:** \tab GPL-3 \cr
 #'   **Zenodo:** \tab \doi{10.5281/zenodo.3526659} \cr
 #'  }
@@ -35,3 +35,6 @@
 #' @docType package
 #' @keywords internal
 "_PACKAGE"
+
+#' @importFrom methods setGeneric setMethod .valueClassTest
+NULL